requirements.txt
# --- ML & data ---
tensorflow-cpu==2.16.1
pandas==2.2.2
pandas_datareader==0.10.0
numpy==1.26.4
scikit-learn==1.5.0
yfinance==0.2.38
ta==0.11.0
joblib==1.4.2

# --- API server ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
